upgrade: Routes moved to App, Links moved to Header; add: Header/Footer, Home components; dirs for every component created
