+++ package.json
@@ -28,7 +28,7 @@
     "history": "4.10.1",
     "jsonwebtoken": "^8.5.1",
     "node": "^14.19.1",
-    "node-fetch": "^3.2.6",
+    "node-fetch": "^2.6.6",
     "prop-types": "15.7.2",
     "ramda": "0.26.1",
     "react": "17.0.1",
