--- server/config.js
@@ -1,7 +1,0 @@
-import dotenv from 'dotenv';
-dotenv.config();
-const port = process.env.PORT;
-const user = process.env.USER;
-const password = process.env.PASSWORD;
-
-export { port, user, password };
