+++ src/components/Error404/Error404.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+type Props = {};
+
+const Error404 = (props: Props): JSX.Element => {
+  return <div>Nothing here!</div>;
+};
+
+export default Error404;
