+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+interface Props {}
+
+const Footer = (props: Props): JSX.Element => {
+  return <div style={{ padding: 20 }}>Footer</div>;
+};
+
+export default Footer;
